first commit of task no JDP-1911-02
