+++ build.gradle
@@ -29,4 +29,6 @@ dependencies {
     implementation('org.springframework.boot:spring-boot-starter-web')
     runtimeOnly('com.h2database:h2')
     testImplementation('org.springframework.boot:spring-boot-starter-test')
+    compile 'org.projectlombok:lombok'
+    annotationProcessor 'org.projectlombok:lombok'
 }
