--- src/main/java/com/kodilla/ecommercee/UserController.java
@@ -1,4 +1,0 @@
-package com.kodilla.ecommercee;
-
-public class UserController {
-}
