+++ src/main/java/com/kodilla/ecommercee/controller/UserController.java
@@ -1,0 +1,4 @@
+package com.kodilla.ecommercee.controller;
+
+public class UserController {
+}
