+++ src/main/java/com/kodilla/ecommercee/mapper/GroupMapper.java
@@ -1,0 +1,11 @@
+package com.kodilla.ecommercee.mapper;
+
+import com.kodilla.ecommercee.domain.GroupEntity;
+import org.springframework.stereotype.Component;
+
+@Component
+public class GroupMapper {
+//    public GroupEntity mapToTask (GroupDto groupDto) {
+//        Map<>
+//    }
+}
